Refactor: `Avatar` 엔티티 관련 클래스 패키지 이동
