+++ src/main/java/com/join/core/avatar/repository/AvatarRepository.java
@@ -1,4 +1,4 @@
-package com.join.core.auth.repository;
+package com.join.core.avatar.repository;
 
 import com.join.core.avatar.domain.Avatar;
 import org.springframework.data.jpa.repository.JpaRepository;
